# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server for an e-commerce application.",
        "title": "Go Ecommerce API",
        "contact": {
            "name": "API Support",
            "email": "niosv.1104@gmail.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/addresses": {
            "get": {
                "description": "Get a paginated list of addresses",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Addresses"
                ],
                "summary": "List addresses",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Items per page",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.AddressResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.StandardResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new address",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Addresses"
                ],
                "summary": "Create a new address",
                "parameters": [
                    {
                        "description": "Address data",
                        "name": "address",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateAddressRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.AddressResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.StandardResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.StandardResponse"
                        }
                    }
                }
            }
        },
        "/addresses/{id}": {
            "get": {
                "description": "Get address details by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Addresses"
                ],
                "summary": "Get address by ID",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Address ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.AddressResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.StandardResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.StandardResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Update an existing address by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Addresses"
                ],
                "summary": "Update an address",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Address ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated address data",
                        "name": "address",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateAddressRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.AddressResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.StandardResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.StandardResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete an address by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Addresses"
                ],
                "summary": "Delete an address",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Address ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.StandardResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.StandardResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.AddressResponse": {
            "type": "object",
            "properties": {
                "city": {
                    "type": "string"
                },
                "country": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_default_billing": {
                    "type": "boolean"
                },
                "is_default_shipping": {
                    "type": "boolean"
                },
                "state": {
                    "type": "string"
                },
                "street": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                },
                "zip_code": {
                    "type": "string"
                }
            }
        },
        "dto.CreateAddressRequest": {
            "type": "object",
            "required": [
                "city",
                "country",
                "state",
                "street",
                "user_id"
            ],
            "properties": {
                "city": {
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 2
                },
                "country": {
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 2
                },
                "is_default_billing": {
                    "type": "boolean"
                },
                "is_default_shipping": {
                    "type": "boolean"
                },
                "state": {
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 2
                },
                "street": {
                    "type": "string",
                    "maxLength": 120,
                    "minLength": 5
                },
                "user_id": {
                    "type": "string"
                },
                "zip_code": {
                    "type": "string",
                    "maxLength": 20
                }
            }
        },
        "dto.UpdateAddressRequest": {
            "type": "object",
            "properties": {
                "city": {
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 2
                },
                "country": {
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 2
                },
                "is_default_billing": {
                    "type": "boolean"
                },
                "is_default_shipping": {
                    "type": "boolean"
                },
                "state": {
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 2
                },
                "street": {
                    "type": "string",
                    "maxLength": 120,
                    "minLength": 5
                },
                "zip_code": {
                    "type": "string",
                    "maxLength": 20
                }
            }
        },
        "response.StandardResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}